DDF-1792 Generify XML transformer
Made XML SaxEventHandler more generic, can now be customized by passing a map of XML Element or attribute value
